parted code for another files
