+++ src/vision/oak/reading_data.py
@@ -146,5 +146,3 @@ def read_pose(pose_path: str) -> np.ndarray:
     if pose.shape == (3, 4):  # Convert 3x4 matrix to 4x4 by adding a row for homogeneous coordinates
         pose = np.vstack((pose, np.array([0, 0, 0, 1], dtype=np.float64)))
     return pose
-
-
